requirements.txt
asgiref==3.7.2
bleach==6.1.0
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloudinary==1.37.0
colorama==0.4.6
coverage==7.3.2
cryptography==41.0.7
defusedxml==0.7.1
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==4.1
django-allauth==0.59.0
django-crispy-forms==1.14.0
django-environ==0.11.2
django-summernote==0.8.20.0
django-taggit==5.0.1
Flask==2.3.0
Flask-SQLAlchemy==2.5.1
google-auth==2.23.2
google-auth-oauthlib==1.1.0
greenlet==3.0.1
gspread==5.11.3
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lolcat==1.4
MarkupSafe==2.1.3
oauth2client==4.1.3
oauthlib==3.2.2
packaging==23.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyfiglet==1.0.2
PyJWT==2.8.0
pyparsing==3.1.1
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
setuptools==69.0.3
six==1.16.0
SQLAlchemy==1.4.46
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
webencodings==0.5.1
Werkzeug==3.0.1
